Modified hyb2onccal to be a callable function.
