+++ isis/src/hayabusa2/apps/hyb2onccal/hyb2onccal.h
@@ -1,0 +1,15 @@
+#ifndef HYB2ONCCAL_H
+#define HYB2ONCCAL_H
+
+#include "UserInterface.h"
+
+namespace Isis {
+
+  extern void hyb2onccal(UserInterface &ui);
+
+}
+
+
+
+
+#endif // HYB2ONCCAL_H
